refactor: uniformise to "blocs"
